Rename consumer to operator
